devtools/client/shared/components/test/mochitest/test_tree_10.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test that when an item in the Tree component is expanded or collapsed the appropriate event handler fires.
-->
<head>
  <meta charset="utf-8">
  <title>Tree component test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <link rel="stylesheet" href="chrome://devtools/skin/light-theme.css" type="text/css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  try {
    const ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    const { createFactory } = browserRequire("devtools/client/shared/vendor/react");
    const Tree = createFactory(browserRequire("devtools/client/shared/components/VirtualizedTree"));

    function renderTree(props) {
      const treeProps = Object.assign({},
        TEST_TREE_INTERFACE,
        { autoExpandDepth: 1 },
        props
      );
      return ReactDOM.render(Tree(treeProps), window.document.body);
    }

    const tree = renderTree({ focused: "A" });

    isRenderedTree(document.body.textContent, [
      "A:true",
      "-B:false",
      "-C:false",
      "-D:false",
      "M:false",
      "-N:false",
    ], "Should have auto-expanded one level.");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
